feat: 이력서 라우팅 구조 설정 및 상태 공유 처리
